requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
newspaper3k>=0.2.8
lxml>=4.9.2  # Required dependency
nltk>=3.5    # Required dependency